// _-同城生活社区平台管理端/src/types.ts
export interface User {
  id: string;
  name: string;
  email: string;
  phone: string;
  avatar: string;
  status: 'normal' | 'disabled';
  verified: 'verified' | 'unverified';
  region: string;
  registerTime: string;
  followersCount: number;
  followingCount: number;
  dynamicsCount: number;
  goodsCount: number;
  servicesCount: number;
}